+++ src/bananagl/bananagl.ts
@@ -12,6 +12,7 @@ import { Renderer } from './renderer/renderer';
 import { mountRenderer, unmountRenderer } from './renderer/setup';
 import { Scene } from './scene/scene';
 import { Selection, SelectionManager } from './scene/selection';
+import { cloneUniforms } from './shaders/clone';
 import { Shader } from './shaders/shader';
 import { Profiler } from './utils/profiler';
 import { Shortcut, ShortcutOnMouseMove, ShortcutOnPress } from './window/shortcuts';
@@ -41,6 +42,7 @@ export {
     ShortcutOnMouseMove,
     mountRenderer,
     unmountRenderer,
+    cloneUniforms,
 };
 
 export type { Selectable, Shortcut };
